REQ6 - Finalizando requisito 6 - [DONE]

+++ desafio1.sql
@@ -5,88 +5,83 @@ CREATE DATABASE SpotifyClone;
 USE SpotifyClone;
 
 CREATE TABLE artistas(
-  artistas_id INT PRIMARY KEY AUTO_INCREMENT,
-  nome VARCHAR(35) NOT NULL
+    artistas_id INT PRIMARY KEY AUTO_INCREMENT,
+    nome VARCHAR(35) NOT NULL
 ) engine = InnoDB;
 
 CREATE TABLE plano(
-  plano_id INT PRIMARY KEY AUTO_INCREMENT,
-  plano VARCHAR(15) NOT NULL,
-  valor DOUBLE NOT NULL
+    plano_id INT PRIMARY KEY AUTO_INCREMENT,
+    plano VARCHAR(15) NOT NULL,
+    valor DECIMAL(4, 2) NOT NULL
 ) engine = InnoDB;
 
 CREATE TABLE usuarios(
-  usuarios_id INT PRIMARY KEY AUTO_INCREMENT,
-  nome VARCHAR(35) NOT NULL,
-  idade INT NOT NULL,
-  plano_id INT NOT NULL,
-  FOREIGN KEY (plano_id) REFERENCES plano (plano_id)
+    usuarios_id INT PRIMARY KEY AUTO_INCREMENT,
+    nome VARCHAR(35) NOT NULL,
+    idade INT NOT NULL,
+    plano_id INT NOT NULL,
+    FOREIGN KEY (plano_id) REFERENCES plano (plano_id)
 ) engine = InnoDB;
 
 CREATE TABLE albuns(
-  albuns_id INT PRIMARY KEY AUTO_INCREMENT,
-  nome VARCHAR(35) NOT NULL,
-  artistas_id INT NOT NULL,
-  FOREIGN KEY (artistas_id) REFERENCES artistas (artistas_id)
+    albuns_id INT PRIMARY KEY AUTO_INCREMENT,
+    nome VARCHAR(35) NOT NULL,
+    artistas_id INT NOT NULL,
+    FOREIGN KEY (artistas_id) REFERENCES artistas (artistas_id)
 ) engine = InnoDB;
 
 CREATE TABLE musicas(
-  musicas_id INT PRIMARY KEY AUTO_INCREMENT,
-  nome VARCHAR(50) NOT NULL,
-  albuns_id INT NOT NULL,
-  FOREIGN KEY (albuns_id) REFERENCES albuns (albuns_id)
+    musicas_id INT PRIMARY KEY AUTO_INCREMENT,
+    nome VARCHAR(50) NOT NULL,
+    albuns_id INT NOT NULL,
+    FOREIGN KEY (albuns_id) REFERENCES albuns (albuns_id)
 ) engine = InnoDB;
 
 CREATE TABLE historico_reproducao(
-  usuarios_id INT NOT NULL,
-  musicas_id INT NOT NULL,
-  CONSTRAINT PRIMARY KEY(usuarios_id, musicas_id),
-  FOREIGN KEY (usuarios_id) REFERENCES usuarios (usuarios_id),
-  FOREIGN KEY (musicas_id) REFERENCES musicas (musicas_id)
+    usuarios_id INT NOT NULL,
+    musicas_id INT NOT NULL,
+    CONSTRAINT PRIMARY KEY(usuarios_id, musicas_id),
+    FOREIGN KEY (usuarios_id) REFERENCES usuarios (usuarios_id),
+    FOREIGN KEY (musicas_id) REFERENCES musicas (musicas_id)
 ) engine = InnoDB;
 
 CREATE TABLE seguindo_artistas(
-  usuarios_id INT NOT NULL,
-  artistas_id INT NOT NULL,
-  CONSTRAINT PRIMARY KEY(usuarios_id, artistas_id),
-  FOREIGN KEY (usuarios_id) REFERENCES usuarios (usuarios_id),
-  FOREIGN KEY (artistas_id) REFERENCES artistas (artistas_id)
+    usuarios_id INT NOT NULL,
+    artistas_id INT NOT NULL,
+	CONSTRAINT PRIMARY KEY(usuarios_id, artistas_id),
+    FOREIGN KEY (usuarios_id) REFERENCES usuarios (usuarios_id),
+    FOREIGN KEY (artistas_id) REFERENCES artistas (artistas_id)
 ) engine = InnoDB;
 
-INSERT INTO
-  artistas (nome)
+INSERT INTO artistas (nome)
 VALUES
   ("Walter Phoenix"),
   ("Peter Strong"),
   ("Lance Day"),
   ("Freedie Shannon");
 
-INSERT INTO
-  plano (plano, valor)
+INSERT INTO plano (plano, valor)
 VALUES
   ("gratuito", 0),
   ("familiar", 7.99),
   ("universitário", 5.99);
 
-INSERT INTO
-  usuarios (nome, idade, plano_id)
+INSERT INTO usuarios (nome, idade, plano_id)
 VALUES
   ("Thati", 23, 1),
   ("Cintia", 35, 2),
   ("Bill", 20, 3),
   ("Roger", 45, 1);
-
-INSERT INTO
-  albuns (nome, artistas_id)
+  
+INSERT INTO albuns (nome, artistas_id)
 VALUES
   ("Envious", 1),
   ("Exuberant", 1),
   ("Hallowed Steam", 2),
   ("Incandescent", 3),
   ("Temporary Culture", 4);
-
-INSERT INTO
-  musicas (nome, albuns_id)
+  
+INSERT INTO musicas (nome, albuns_id)
 VALUES
   ('Soul For Us', 1),
   ('Reflections Of Magic', 1),
@@ -106,10 +101,9 @@ VALUES
   ('Thang Of Thunder', 5),
   ('Words Of Her Life', 5),
   ('Without My Streets', 5);
-
-INSERT INTO
-  historico_reproducao (usuarios_id, musicas_id)
-VALUES
+  
+  INSERT INTO historico_reproducao (usuarios_id, musicas_id)
+  VALUES
   (1, 1),
   (1, 6),
   (1, 14),
@@ -124,10 +118,9 @@ VALUES
   (4, 3),
   (4, 18),
   (4, 11);
-
-INSERT INTO
-  seguindo_artistas (usuarios_id, artistas_id)
-VALUES
+  
+  INSERT INTO seguindo_artistas (usuarios_id, artistas_id)
+  VALUES
   (1, 1),
   (1, 4),
   (1, 3),
@@ -136,3 +129,4 @@ VALUES
   (3, 2),
   (3, 1),
   (4, 4);
+
